Added mouse input events
This includes mouse pressed, mouse released, and mouse moved. Mouse
events are different from keyboard events in that they also have the x
and y position of the cursor, which is translated into the
game's coordinate system.

Moving from the window's coordinate system to the game's coordinates
had to go via the renderer, since it knows how to relate the two to
each other.